readonly

+++ DB.py
@@ -1,14 +1,15 @@
 import json
 
 class DB:
-    def __init__(self, dbPath):
+    def __init__(self, dbPath,readOnly=False):
         self.path = dbPath
         f = open(self.path).read()
         self.db = json.loads(f)
 
     def __del__(self):
-        with open(self.path, "w") as outfile: 
-            json.dump(self.db, outfile)
+        if not self.readOnly:
+            with open(self.path, "w") as outfile: 
+                json.dump(self.db, outfile)
 
     def set(self,k,v):
         self.db[k] = v
